moved/renamed player.ts to character/index.ts, separated non common states and commons into new character.
