+++ phaser/src/editor/components/box/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as _ from 'lodash';
-import { BoxConfig, BoxType, isCircleBox } from 'src/characters';
+import { BoxConfig, BoxType, isCircleBox } from 'src/characters/frameData';
 import cx from 'classnames';
 import 'src/editor/components/box/styles.scss';
 import { Vector2 } from '@lawsumisu/common-utilities';
